Converting code to run under Python 3 (retaining Python2 compatability)
The Python code now needs the 'six' package to provide a smoother
interface to text/binary strings
	(easy_install six, port install six, pip install six, whichever works for you)

In extracttable.py, a few bare strings were converted to binary string with six.b(CHAR) as needed. The
Regexp was converted to a binary string. A comparison was changed to look at the string
length, not its equality with "". Text is decoded to UTF-8 before writing to XML.

In mkinc.py and periphOfTab.py, text is decoded to utf-8.